requirements.txt
pillow
synthizer